24-05-2019 Part 1 (with some bug fixes)

+++ Done/ft_strnstr.c
@@ -6,15 +6,17 @@
 /*   By: jpieczar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 14:31:50 by jpieczar          #+#    #+#             */
-/*   Updated: 2019/05/22 09:14:44 by jpieczar         ###   ########.fr       */
+/*   Updated: 2019/05/24 10:47:48 by jpieczar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 char    *ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
     size_t	i;
-    int		j;
-    int		p;
+    size_t	j;
+    size_t	p;
 
     i = 0;
     j = 0;
